add fetching contract info to account/detail.js
